+++ src/main/java/Ex03_08.java
@@ -1,7 +1,7 @@
 public class Ex03_08 {
     public static void main(String[] args) {
         int a,b;
-        float c,d;
+        double c,d;
 
         a = 100;
         b = a;
