+++ src/main/java/Ex03_10.java
@@ -1,7 +1,7 @@
 public class Ex03_10 {
     public static void main(String[] args) {
         int a = 100, b = 200;
-        float result;
+        double result;
 
         result = a/ b;
